Move iterator.rs, add type_from_str impl for iterators
